# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bricks.bio API
  description: |
    Universal API for retrieving biological parts from multiple providers. 
    We provide a new standardized JSON schema (called biobrick). 
    Our goal is to aggregate data to provide a modern and unified tool to accelerate synthetic biology evolution, especially towards iterative discovery and automation.
    ## Providers

    - [`iGEM Parts Registry (Legacy)`](https://parts.igem.org/) is a HTML scraper.
    - [`iGEM Registry`](https://registry.igem.org) is a single JSON request.
    - [`iGEM via SynBioHub`](https://synbiohub.org/public/igem/igem_collection/1) is a SBOL and GenBank parser.
    - [`Ensembl`](https://www.ensembl.org) is a GenBank parser.
    - [`NCBI`](https://www.ncbi.nlm.nih.gov/) is a GenBank parser.
    - [`AddGene`](https://www.addgene.org/) is a HTML scraper.

    ## Types

      - `canonical` is a standardized custom format.
      - `ontology` is the [Sequence Ontology](http://sequenceontology.org/) identifier.
      - `visual` is the type's name in [SBOL Visual CSS](https://mtripnaux.github.io/SBOL-Visual-CSS/reference.html).
      - `slug` is the new iGEM standard for storing types.
  version: 0.1.0
  licence: GPL-3.0+
  contact:
    name: Matheo Tripnaux
    url: https://matheo.tripnaux.com
    email: matheo@tripnaux.com
servers:
  - url: https://bricks.bio
    description: Production server
paths:
  /parts/{id}:
    get:
      summary: Get biobrick by ID
      description: Searches for a part by ID across multiple providers (have a look at the full list on GitHub) and returns enriched metadata, sequence, and features.
      parameters:
        - name: id
          in: path
          required: true
          example: BBa_B0034
          schema:
            type: string
      responses:
        '200':
          description: Successful response in Biobrick format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Biobrick'
        '404':
          description: Part not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /parts/{id}/sbol:
    get:
      summary: Get biobrick in SBOL
      description: Searches for a part by ID and returns enriched data in Synthetic Biology Open Language (SBOL v2.0) format. Note that not all Biobricks metadata may be represented in SBOL.
      parameters:
        - name: id
          in: path
          required: true
          example: BBa_B0034
          schema:
            type: string
      responses:
        '200':
          description: Successful response in SBOL format
          content:
            application/rdf+xml:
              schema:
                type: string
                format: xml
        '404':
          description: Part not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cache/search:
    get:
      summary: Search cached biobricks
      description: Performs semantic search over the locally cached biobricks and returns the most relevant matches.
      parameters:
        - name: q
          in: query
          required: true
          description: Search query string.
          example: insulin promoter
          schema:
            type: string
        - name: n
          in: query
          required: false
          description: Maximum number of results to return. Values above 50 are capped to 50.
          example: 10
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Ranked cached results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Search'
        '400':
          description: Missing or empty query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Cache search failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cache/stats:
    get:
      summary: Get cache statistics
      description: Returns the number of biobrick entries currently stored in the local cache.
      responses:
        '200':
          description: Current cache statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheStats'
        '500':
          description: Cache statistics could not be read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stats:
    get:
      summary: Get API usage stats
      description: Returns the number of unique clients and total requests over various time periods.
      responses:
        '200':
          description: Current API usage statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiStats'
        '500':
          description: Statistics could not be read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ontology:
    get:
      summary: Get ontology data
      description: Returns a list of types and their matching with different ontologies.
      responses:
        '200':
          description: Current ontology data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Type'
components:
  schemas:
    ApiStats:
      type: object
      required:
        - unique_ips_today
        - unique_ips_total
        - requests_today
        - requests_total
        - unique_ips_last_7_days
      properties:
        unique_ips_today:
          type: integer
          example: 12
        unique_ips_total:
          type: integer
          example: 340
        requests_today:
          type: integer
          example: 87
        requests_total:
          type: integer
          example: 15420
        unique_ips_last_7_days:
          type: integer
          example: 64
    CacheStats:
      type: object
      required:
        - entries
      properties:
        entries:
          type: integer
          example: 42
    Search:
      type: object
      required:
        - query
        - requested
        - count
        - results
      properties:
        query:
          type: string
          example: insulin promoter
        requested:
          type: integer
          example: 10
        count:
          type: integer
          example: 3
        results:
          type: array
          items:
            $ref: '#/components/schemas/Hit'
    Hit:
      type: object
      required:
        - match
        - biobrick
      properties:
        match:
          type: number
          format: double
          example: 0.8731
        biobrick:
          $ref: '#/components/schemas/Biobrick'
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Part not found
    Biobrick:
      type: object
      required:
        - metadata
        - sequence
        - features
      properties:
        metadata:
          $ref: '#/components/schemas/MetaBiobrick'
        sequence:
          type: string
          description: The nucleotide sequence of the part.
        features:
          type: array
          items:
            $ref: '#/components/schemas/MetaFeature'
    MetaBiobrick:
      type: object
      required:
        - id
        - name
        - description
        - creation
        - size
        - circular
        - type
        - authors
        - providers
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        creation:
          type: string
          format: date-time
          description: ISO 8601 creation date.
        size:
          type: integer
        circular:
          type: boolean
        type:
          $ref: '#/components/schemas/Type'
        authors:
          type: array
          items:
            $ref: '#/components/schemas/Author'
        providers:
          type: array
          items:
            $ref: '#/components/schemas/MetaProvider'
    Author:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        role:
          type: string
          nullable: true
    MetaProvider:
      type: object
      required:
        - name
        - link
        - date
      properties:
        name:
          type: string
        link:
          type: string
        date:
          type: string
          format: date-time
    MetaFeature:
      type: object
      required:
        - id
        - name
        - type
        - location
      properties:
        id:
          type: string
        name:
          type: string
        type:
          $ref: '#/components/schemas/Type'
        location:
          $ref: '#/components/schemas/Location'
    Location:
      type: object
      required:
        - start
        - end
        - forward
      properties:
        start:
          type: integer
        end:
          type: integer
        forward:
          type: boolean
          description: "Whether the feature is on the forward strand."
    Type:

        type: object
        description: Represents the type of a sequence.
        required:
          - canonical
          - visual
          - also
          - slug
        properties:
          canonical:
            type: string
            description: Standardized custom format.
          ontology:
            type:
              - string
              - 'null'
            description: Sequence Ontology identifier.
          visual:
            type: 
              - string
              - 'null'
            description: Type class name in SBOL Visual CSS.
          also:
            type: array
            description: Alternative names and synonyms used to infer the type.
            items:
              type: string
          slug:
            type: 
              - string
              - 'null'
            description: New iGEM standard for storing types.